+++ relm4/src/runtime_util.rs
@@ -78,7 +78,7 @@ impl<Output, Command> RuntimeSenders<Output, Command> {
         // Notifies the component's child commands that it is being shut down.
         let (shutdown_notifier, shutdown_recipient) = shutdown::channel();
 
-        // Cannel to tell the component to shutdown.
+        // Channel to tell the component to shutdown.
         // Use a capacity of 2 to prevent blocking when the runtime is dropped while app shutdown is running as well.
         // This rare case will emit 2 messages (but certainly not more).
         let (shutdown_event_sender, shutdown_event_receiver) = mpsc::channel(2);